keep shared request handling together
