+++ pkg/coordinator/engine/vllm/helpers.go
@@ -17,10 +17,6 @@ limitations under the License.
 package vllm
 
 import (
-	"encoding/json"
-	"fmt"
-	"math"
-
 	reqcommon "github.com/llm-d/llm-d-router/pkg/common/request"
 	"github.com/llm-d/llm-d-router/pkg/coordinator/gateway"
 	"github.com/llm-d/llm-d-router/pkg/coordinator/pipeline"
@@ -133,250 +129,4 @@ func setGenerateTransferParams(sampling map[string]any, kvParams any, ecParams m
 	if len(ecParams) > 0 {
 		extraArgs[reqcommon.FieldECTransferParams] = ecParams
 	}
-}
-
-// extractTokenIDs converts body["token_ids"] from a JSON-unmarshalled value to []int.
-// Returns ErrBadRequest when the field is absent, not an array, empty, or contains
-// non-integer or negative values.
-func extractTokenIDs(raw any) ([]int, error) {
-	if raw == nil {
-		return nil, fmt.Errorf("token_ids is required: %w", pipeline.ErrBadRequest)
-	}
-	arr, ok := raw.([]any)
-	if !ok {
-		return nil, fmt.Errorf("token_ids must be an array, got %T: %w", raw, pipeline.ErrBadRequest)
-	}
-	if len(arr) == 0 {
-		return nil, fmt.Errorf("token_ids must not be empty: %w", pipeline.ErrBadRequest)
-	}
-	return toIntSlice(arr)
-}
-
-// mmImageArray reads features[field][modalityImage] as a JSON array. present is
-// false when field or its image entry is absent or null, a valid "no such
-// modality" state rather than an error. A present value of the wrong type
-// (field not an object, or the image entry not an array) is ErrBadRequest, so a
-// malformed request fails loudly instead of being silently coerced to absent.
-func mmImageArray(features map[string]any, field string) (arr []any, present bool, err error) {
-	rawField, ok := features[field]
-	if !ok || rawField == nil {
-		return nil, false, nil
-	}
-	m, ok := rawField.(map[string]any)
-	if !ok {
-		return nil, false, fmt.Errorf("%s must be an object: %w", field, pipeline.ErrBadRequest)
-	}
-	rawImage, ok := m[modalityImage]
-	if !ok || rawImage == nil {
-		return nil, false, nil
-	}
-	arr, ok = rawImage.([]any)
-	if !ok {
-		return nil, false, fmt.Errorf("%s[%s] must be an array: %w", field, modalityImage, pipeline.ErrBadRequest)
-	}
-	return arr, true, nil
-}
-
-// extractMultimodalEntries builds []pipeline.MultimodalEntry from the parallel
-// slices in a generate-format features map. Returns nil when features is nil or
-// mm_hashes.image is absent or empty (text-only request).
-//
-// mm_hashes and mm_placeholders are required and must be the same length.
-// kwargs_data is optional: an absent field means every item resolves from the
-// encoder cache by hash (a cache-hit request), so each entry's KwargsData is "".
-// When present, kwargs_data must be parallel to mm_hashes, but an individual
-// item may be null (a cache hit within a mixed batch), which maps to "".
-//
-// Returns ErrBadRequest when a present field has the wrong type, required slices
-// have different lengths, or any element has an unexpected type.
-func extractMultimodalEntries(features map[string]any) ([]pipeline.MultimodalEntry, error) {
-	if features == nil {
-		return nil, nil
-	}
-	rawHashes, _, err := mmImageArray(features, "mm_hashes")
-	if err != nil {
-		return nil, err
-	}
-	// mmImageArray returns a nil slice whenever the field is absent, so the
-	// length check subsumes the presence check here.
-	if len(rawHashes) == 0 {
-		return nil, nil
-	}
-
-	rawPlaceholders, present, err := mmImageArray(features, "mm_placeholders")
-	if err != nil {
-		return nil, err
-	}
-	if !present {
-		return nil, fmt.Errorf("mm_placeholders[%s] is required when mm_hashes[%s] is set: %w",
-			modalityImage, modalityImage, pipeline.ErrBadRequest)
-	}
-
-	rawKwargs, hasKwargs, err := mmImageArray(features, "kwargs_data")
-	if err != nil {
-		return nil, err
-	}
-
-	n := len(rawHashes)
-	if len(rawPlaceholders) != n {
-		return nil, fmt.Errorf("features length mismatch: mm_hashes has %d, mm_placeholders has %d: %w",
-			n, len(rawPlaceholders), pipeline.ErrBadRequest)
-	}
-	// When present, kwargs_data is parallel to mm_hashes: full length with null
-	// placeholders for cached items, never a shortened list. The whole field is
-	// absent for metadata-only (cache-hit) requests.
-	if hasKwargs && len(rawKwargs) != n {
-		return nil, fmt.Errorf("features length mismatch: mm_hashes has %d, kwargs_data has %d: %w",
-			n, len(rawKwargs), pipeline.ErrBadRequest)
-	}
-
-	entries := make([]pipeline.MultimodalEntry, n)
-	for i := range entries {
-		hash, ok := rawHashes[i].(string)
-		if !ok {
-			return nil, fmt.Errorf("mm_hashes[%d] must be a string: %w", i, pipeline.ErrBadRequest)
-		}
-
-		pMap, ok := rawPlaceholders[i].(map[string]any)
-		if !ok {
-			return nil, fmt.Errorf("mm_placeholders[%d] must be an object: %w", i, pipeline.ErrBadRequest)
-		}
-		// The non-negative guarantee here is load-bearing, not just input
-		// hygiene. Engine.buildEncodeTokenIDs indexes fullTokenIDs[offset]
-		// (guarded only on the upper bound) and allocates make([]int, 1+length);
-		// a negative offset or length panics there. vLLM's own schema declares
-		// these as plain ints and accepts negatives, so this stays stricter
-		// deliberately. Do not relax it to a plain int parse.
-		offset, err := anyToNonNegativeInt(pMap["offset"])
-		if err != nil {
-			return nil, fmt.Errorf("mm_placeholders[%d].offset: %v: %w", i, err, pipeline.ErrBadRequest)
-		}
-		length, err := anyToNonNegativeInt(pMap["length"])
-		if err != nil {
-			return nil, fmt.Errorf("mm_placeholders[%d].length: %v: %w", i, err, pipeline.ErrBadRequest)
-		}
-
-		// Empty KwargsData is the sentinel for "resolve from cache": either the
-		// whole kwargs_data field is absent or this item is null.
-		var kwarg string
-		if hasKwargs {
-			switch k := rawKwargs[i].(type) {
-			case string:
-				kwarg = k
-			case nil:
-			default:
-				return nil, fmt.Errorf("kwargs_data[%d] must be a string or null: %w", i, pipeline.ErrBadRequest)
-			}
-		}
-
-		entries[i] = pipeline.MultimodalEntry{
-			Index:      i,
-			Hash:       hash,
-			KwargsData: kwarg,
-			Placeholder: pipeline.PlaceholderRange{
-				Offset: offset,
-				Length: length,
-			},
-		}
-	}
-	return entries, nil
-}
-
-// validateSamplingParams checks that sampling_params and its nested extra_args,
-// when present, are JSON objects. Both are optional. The decode step merges
-// kv_transfer_params into sampling_params.extra_args; a non-object at either
-// level would fall into its fallback branch and be silently replaced with an
-// empty map, discarding client-requested generation parameters with no error.
-// Validating once at ingestion keeps that path fail-loud, consistent with
-// token_ids and features.
-func validateSamplingParams(body map[string]any) error {
-	raw, ok := body[reqcommon.FieldSamplingParams]
-	if !ok || raw == nil {
-		return nil
-	}
-	sampling, ok := raw.(map[string]any)
-	if !ok {
-		return fmt.Errorf("%s must be an object, got %T: %w",
-			reqcommon.FieldSamplingParams, raw, pipeline.ErrBadRequest)
-	}
-	ea, ok := sampling[reqcommon.FieldExtraArgs]
-	if !ok || ea == nil {
-		return nil
-	}
-	if _, ok := ea.(map[string]any); !ok {
-		return fmt.Errorf("%s.%s must be an object, got %T: %w",
-			reqcommon.FieldSamplingParams, reqcommon.FieldExtraArgs, ea, pipeline.ErrBadRequest)
-	}
-	return nil
-}
-
-// toIntSlice converts a JSON-unmarshalled []any of numeric elements to []int.
-// Each element must be a non-negative integer represented as float64 or json.Number.
-// The returned error identifies the offending element by index and wraps
-// pipeline.ErrBadRequest.
-func toIntSlice(values []any) ([]int, error) {
-	out := make([]int, 0, len(values))
-	for i, v := range values {
-		n, err := anyToNonNegativeInt(v)
-		if err != nil {
-			return nil, fmt.Errorf("invalid token at index %d: %v: %w", i, err, pipeline.ErrBadRequest)
-		}
-		out = append(out, n)
-	}
-	return out, nil
-}
-
-// anyToNonNegativeInt converts a single JSON-unmarshalled numeric value to a non-negative int.
-func anyToNonNegativeInt(v any) (int, error) {
-	switch n := v.(type) {
-	case float64:
-		if n < 0 || n != math.Trunc(n) {
-			return 0, fmt.Errorf("expected non-negative integer, got %v", v)
-		}
-		// An in-range integer-valued float64 round-trips through int; a value
-		// too large to fit does not (the conversion saturates), so this rejects
-		// overflow without depending on the fragile float64(MaxInt) boundary.
-		i := int(n)
-		if float64(i) != n {
-			return 0, fmt.Errorf("expected non-negative integer, got %v", v)
-		}
-		return i, nil
-	case json.Number:
-		i, err := n.Int64()
-		if err != nil {
-			return 0, err
-		}
-		if i < 0 || i > math.MaxInt {
-			return 0, fmt.Errorf("expected non-negative integer, got %d", i)
-		}
-		return int(i), nil
-	default:
-		return 0, fmt.Errorf("expected number, got %T", v)
-	}
-}
-
-// validatePlaceholderBounds checks that every placeholder span [offset,
-// offset+length) lies within a prompt of tokenCount tokens. It guards the
-// generate path, where the client supplies placeholder geometry directly:
-// Engine.buildEncodeTokenIDs indexes token_ids[offset] and allocates
-// make([]int, 1+length), so an out-of-range offset reads the wrong token and an
-// unbounded length (a tiny request can claim billions) is a memory-exhaustion
-// vector. vLLM declares offset/length as plain unbounded ints on the generate
-// endpoint and does not enforce this, so the coordinator does. offset and
-// length are already guaranteed non-negative by extractMultimodalEntries.
-func validatePlaceholderBounds(entries []pipeline.MultimodalEntry, tokenCount int) error {
-	for i, e := range entries {
-		off := e.Placeholder.Offset
-		length := e.Placeholder.Length
-		if off >= tokenCount {
-			return fmt.Errorf("mm_placeholders[%d].offset %d out of range for %d token_ids: %w",
-				i, off, tokenCount, pipeline.ErrBadRequest)
-		}
-		// off < tokenCount, so tokenCount-off is positive and cannot overflow.
-		if length > tokenCount-off {
-			return fmt.Errorf("mm_placeholders[%d] span (offset %d + length %d) exceeds %d token_ids: %w",
-				i, off, length, tokenCount, pipeline.ErrBadRequest)
-		}
-	}
-	return nil
 }
